+++ src/scenes/BootScene.js
@@ -15,6 +15,10 @@ export default class BootScene extends Phaser.Scene {
     this.load.image('slime001_reaction',  'Images/slime001_reaction.png');
     this.load.image('slime001_attack',    'Images/slime001_attack.png');
     this.load.image('slime001_dead',      'Images/slime001_dead.png');
+    this.load.image('bg002_rear',   'Images/background002_rear.png');
+    this.load.image('bg002_mid',    'Images/background002_mid.png');
+    this.load.image('bg002_front',  'Images/background002_front.png');
+    this.load.image('ground001',    'Images/ground001.png');
     this.load.image('sewerrat001_default',  'Images/sewerrat001_default.png');
     this.load.image('sewerrat001_reaction', 'Images/sewerrat001_reaction.png');
     this.load.image('sewerrat001_attack',   'Images/sewerrat001_attack.png');